Changes Added

+++ application/helpers/paytm_helper.php
@@ -43,7 +43,7 @@ function pkcs5_unpad_e($text) {
 
 function generateSalt_e($length) {
 	$random = "";
-	srand((double) microtime() * 1000000);
+	srand((float) microtime() * 1000000);
 
 	$data = "AbcDE123IJKLMN67QRSTUVWXYZ";
 	$data .= "aBCdefghijklmn123opq45rs67tuv89wxyz";
